+++ packages/stream_chat_flutter_core/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 3.4.0
+- Updated `stream_chat` dependency to [`3.4.0`](https://pub.dev/packages/stream_chat/changelog).
 
 🐞 Fixed